@@ -5,40 +5,44 @@ members = [
     "icx-cert",
     "ic-identity-hsm",
     "ic-utils",
+    "ic-utils-bindgen",
     "ic-transport-types",
     "icx",
     "ref-tests",
+    "ic-utils-bindgen-tests",
 ]
 
 [workspace.package]
-version = "0.40.1"
+version = "0.45.0"
 authors = ["DFINITY Stiftung <sdk@dfinity.org>"]
 edition = "2021"
 repository = "https://github.com/dfinity/agent-rs"
 # MSRV
 # Avoid updating this field unless we use new Rust features
 # Sync rust-version in rust-toolchain.toml
-rust-version = "1.78.0"
+rust-version = "1.88.0"
 license = "Apache-2.0"
 
 [workspace.lints.clippy]
 needless_lifetimes = "allow"
 
 [workspace.dependencies]
-ic-agent = { path = "ic-agent", version = "0.40.1", default-features = false }
-ic-utils = { path = "ic-utils", version = "0.40.1" }
-ic-transport-types = { path = "ic-transport-types", version = "0.40.1" }
+ic-agent = { path = "ic-agent", version = "0.45.0", default-features = false }
+ic-utils = { path = "ic-utils", version = "0.45.0" }
+ic-transport-types = { path = "ic-transport-types", version = "0.45.0" }
+ref-tests = { path = "ref-tests" }
 
 ic-certification = "3"
+anyhow = { version = "1.0", features = ["backtrace"] }
 candid = "0.10.10"
-candid_parser = "0.1.4"
+candid_parser = "0.2.2"
 clap = "4.5.21"
-ed25519-consensus = "2.1.0"
 futures-util = "0.3.31"
 hex = "0.4.3"
+ic-ed25519 = "0.5"
 k256 = "0.13.4"
 leb128 = "0.2.5"
-pocket-ic = "6.0.0"
+pocket-ic = "10.0.0"
 p256 = "0.13.2"
 rand = "0.8.5"
 reqwest = { version = "0.12", default-features = false }
@@ -51,3 +55,4 @@ sha2 = "0.10.8"
 thiserror = "2.0.3"
 time = "0.3"
 tokio = { version = "1.41.1", default-features = false }
+ic-management-canister-types = "0.5.0"
