@@ -10,6 +10,7 @@ pub(crate) mod nonce;
 pub(crate) mod response_authentication;
 pub mod route_provider;
 pub mod status;
+pub mod subnet;
 
 pub use agent_config::AgentConfig;
 pub use agent_error::AgentError;
@@ -17,25 +18,23 @@ use agent_error::{HttpErrorPayload, Operation};
 use async_lock::Semaphore;
 use async_trait::async_trait;
 pub use builder::AgentBuilder;
+use bytes::Bytes;
 use cached::{Cached, TimedCache};
-use ed25519_consensus::{Error as Ed25519Error, Signature, VerificationKey};
-use futures_util::StreamExt;
-use http::{header::CONTENT_TYPE, HeaderMap, Method, StatusCode};
+use http::{header::CONTENT_TYPE, HeaderMap, Method, StatusCode, Uri};
+use ic_ed25519::{PublicKey, SignatureError};
 #[doc(inline)]
 pub use ic_transport_types::{
     signed, CallResponse, Envelope, EnvelopeContent, RejectCode, RejectResponse, ReplyResponse,
     RequestStatusResponse,
 };
 pub use nonce::{NonceFactory, NonceGenerator};
-use rangemap::{RangeInclusiveMap, RangeInclusiveSet, StepFns};
-use reqwest::{Body, Client, Request, Response};
+use rangemap::{RangeInclusiveMap, StepFns};
+use reqwest::{Client, Request, Response};
 use route_provider::{
-    dynamic_routing::{
-        dynamic_route_provider::DynamicRouteProviderBuilder, node::Node,
-        snapshot::latency_based_routing::LatencyRoutingSnapshot,
-    },
+    dynamic_routing::{dynamic_route_provider::DynamicRouteProviderBuilder, node::Node},
     RouteProvider, UrlUntilReady,
 };
+pub use subnet::Subnet;
 use time::OffsetDateTime;
 use tower_service::Service;
 
@@ -44,9 +43,12 @@ mod agent_test;
 
 use crate::{
     agent::response_authentication::{
-        extract_der, lookup_canister_info, lookup_canister_metadata, lookup_request_status,
-        lookup_subnet, lookup_subnet_metrics, lookup_time, lookup_value,
+        extract_der, lookup_canister_info, lookup_canister_metadata, lookup_canister_ranges,
+        lookup_incomplete_subnet, lookup_request_status, lookup_subnet_and_ranges,
+        lookup_subnet_canister_ranges, lookup_subnet_metrics, lookup_time, lookup_tree,
+        lookup_value,
     },
+    agent_error::TransportError,
     export::Principal,
     identity::Identity,
     to_request_id, RequestId,
@@ -62,11 +64,11 @@ use serde::Serialize;
 use status::Status;
 use std::{
     borrow::Cow,
-    collections::HashMap,
     convert::TryFrom,
     fmt::{self, Debug},
     future::{Future, IntoFuture},
     pin::Pin,
+    str::FromStr,
     sync::{Arc, Mutex, RwLock},
     task::{Context, Poll},
     time::Duration,
@@ -86,9 +88,8 @@ type AgentFuture<'a, V> = Pin<Box<dyn Future<Output = Result<V, AgentError>> + '
 
 /// A low level Agent to make calls to a Replica endpoint.
 ///
-/// ```ignore
-/// # // This test is ignored because it requires an ic to be running. We run these
-/// # // in the ic-ref workflow.
+#[cfg_attr(unix, doc = " ```rust")] // pocket-ic
+#[cfg_attr(not(unix), doc = " ```ignore")]
 /// use ic_agent::{Agent, export::Principal};
 /// use candid::{Encode, Decode, CandidType, Nat};
 /// use serde::Deserialize;
@@ -104,14 +105,13 @@ type AgentFuture<'a, V> = Pin<Box<dyn Future<Output = Result<V, AgentError>> + '
 /// }
 ///
 /// # fn create_identity() -> impl ic_agent::Identity {
-/// #
-/// #     ic_agent::identity::BasicIdentity::from_signing_key(
-/// #         ed25519_consensus::SigningKey::new(rand::thread_rng())
-/// #     )
+/// #     // In real code, the raw key should be either read from a pem file or generated with randomness.
+/// #     ic_agent::identity::BasicIdentity::from_raw_key(&[0u8;32])
 /// # }
 /// #
 /// async fn create_a_canister() -> Result<Principal, Box<dyn std::error::Error>> {
-/// # let url = format!("http://localhost:{}", option_env!("IC_REF_PORT").unwrap_or("4943"));
+/// # Ok(ref_tests::utils::with_pic(async move |pic| {
+/// # let url = ref_tests::utils::get_pic_url(&pic);
 ///   let agent = Agent::builder()
 ///     .with_url(url)
 ///     .with_identity(create_identity())
@@ -124,10 +124,10 @@ type AgentFuture<'a, V> = Pin<Box<dyn Future<Output = Result<V, AgentError>> + '
 ///   agent.fetch_root_key().await?;
 ///   let management_canister_id = Principal::from_text("aaaaa-aa")?;
 ///
-///   // Create a call to the management canister to create a new canister ID,
-///   // and wait for a result.
-///   // The effective canister id must belong to the canister ranges of the subnet at which the canister is created.
-///   let effective_canister_id = Principal::from_text("rwlgt-iiaaa-aaaaa-aaaaa-cai").unwrap();
+///   // Create a call to the management canister to create a new canister ID, and wait for a result.
+///   // This API only works in local instances; mainnet instances must use the cycles ledger.
+///   // See `dfx info default-effective-canister-id`.
+/// # let effective_canister_id = ref_tests::utils::get_effective_canister_id(&pic).await;
 ///   let response = agent.update(&management_canister_id, "provisional_create_canister_with_cycles")
 ///     .with_effective_canister_id(effective_canister_id)
 ///     .with_arg(Encode!(&Argument { amount: None })?)
@@ -136,6 +136,7 @@ type AgentFuture<'a, V> = Pin<Box<dyn Future<Output = Result<V, AgentError>> + '
 ///   let result = Decode!(response.as_slice(), CreateCanisterResult)?;
 ///   let canister_id: Principal = Principal::from_text(&result.canister_id.to_text())?;
 ///   Ok(canister_id)
+/// # }).await)
 /// }
 ///
 /// # let mut runtime = tokio::runtime::Runtime::new().unwrap();
@@ -217,13 +218,9 @@ impl Agent {
                     );
                     let seeds = vec![Node::new(url.domain().unwrap()).unwrap()];
                     UrlUntilReady::new(url, async move {
-                        DynamicRouteProviderBuilder::new(
-                            LatencyRoutingSnapshot::new(),
-                            seeds,
-                            client,
-                        )
-                        .build()
-                        .await
+                        let provider = DynamicRouteProviderBuilder::new(seeds, client).build();
+                        provider.start().await;
+                        provider
                     }) as Arc<dyn RouteProvider>
                 } else {
                     Arc::new(url)
@@ -320,7 +317,7 @@ impl Agent {
         let bytes = self
             .execute(
                 Method::POST,
-                &format!("api/v2/canister/{}/query", effective_canister_id.to_text()),
+                &format!("api/v3/canister/{}/query", effective_canister_id.to_text()),
                 Some(serialized_bytes),
             )
             .await?
@@ -338,7 +335,7 @@ impl Agent {
     {
         let _permit = self.concurrent_requests_semaphore.acquire().await;
         let endpoint = format!(
-            "api/v2/canister/{}/read_state",
+            "api/v3/canister/{}/read_state",
             effective_canister_id.to_text()
         );
         let bytes = self
@@ -357,7 +354,7 @@ impl Agent {
         A: serde::de::DeserializeOwned,
     {
         let _permit = self.concurrent_requests_semaphore.acquire().await;
-        let endpoint = format!("api/v2/subnet/{}/read_state", subnet_id.to_text());
+        let endpoint = format!("api/v3/subnet/{}/read_state", subnet_id.to_text());
         let bytes = self
             .execute(Method::POST, &endpoint, Some(serialized_bytes))
             .await?
@@ -371,7 +368,7 @@ impl Agent {
         serialized_bytes: Vec<u8>,
     ) -> Result<TransportCallResponse, AgentError> {
         let _permit = self.concurrent_requests_semaphore.acquire().await;
-        let endpoint = format!("api/v3/canister/{}/call", effective_canister_id.to_text());
+        let endpoint = format!("api/v4/canister/{}/call", effective_canister_id.to_text());
         let (status_code, response_body) = self
             .execute(Method::POST, &endpoint, Some(serialized_bytes))
             .await?;
@@ -515,25 +512,17 @@ impl Agent {
                         actual: node_key[..12].to_vec(),
                     });
                 }
-                let pubkey =
-                    VerificationKey::try_from(<[u8; 32]>::try_from(&node_key[12..]).unwrap())
-                        .map_err(|_| AgentError::MalformedPublicKey)?;
-                let sig = Signature::from(
-                    <[u8; 64]>::try_from(&signature.signature[..])
-                        .map_err(|_| AgentError::MalformedSignature)?,
-                );
+                let pubkey = PublicKey::deserialize_raw(&node_key[12..])
+                    .map_err(|_| AgentError::MalformedPublicKey)?;
 
-                match pubkey.verify(&sig, &signable) {
-                    Err(Ed25519Error::InvalidSignature) => {
+                match pubkey.verify_signature(&signable, &signature.signature[..]) {
+                    Ok(()) => (),
+                    Err(SignatureError::InvalidSignature) => {
                         return Err(AgentError::QuerySignatureVerificationFailed)
                     }
-                    Err(Ed25519Error::InvalidSliceLength) => {
+                    Err(SignatureError::InvalidLength) => {
                         return Err(AgentError::MalformedSignature)
                     }
-                    Err(Ed25519Error::MalformedPublicKey) => {
-                        return Err(AgentError::MalformedPublicKey)
-                    }
-                    Ok(()) => (),
                     _ => unreachable!(),
                 }
             }
@@ -732,16 +721,21 @@ impl Agent {
         let mut retry_policy = self.get_retry_policy();
 
         let mut request_accepted = false;
-        let (resp, cert) = self
-            .request_status_signed(
-                request_id,
-                effective_canister_id,
-                signed_request_status.clone(),
-            )
-            .await?;
         loop {
+            let (resp, cert) = self
+                .request_status_signed(
+                    request_id,
+                    effective_canister_id,
+                    signed_request_status.clone(),
+                )
+                .await?;
             match resp {
-                RequestStatusResponse::Unknown => {}
+                RequestStatusResponse::Unknown => {
+                    // If status is still `Unknown` after 5 minutes, the ingress message is lost.
+                    if retry_policy.get_elapsed_time() > Duration::from_secs(5 * 60) {
+                        return Err(AgentError::TimeoutWaitingForResponse());
+                    }
+                }
 
                 RequestStatusResponse::Received | RequestStatusResponse::Processing => {
                     if !request_accepted {
@@ -800,7 +794,12 @@ impl Agent {
                 .request_status_raw(request_id, effective_canister_id)
                 .await?;
             match resp {
-                RequestStatusResponse::Unknown => {}
+                RequestStatusResponse::Unknown => {
+                    // If status is still `Unknown` after 5 minutes, the ingress message is lost.
+                    if retry_policy.get_elapsed_time() > Duration::from_secs(5 * 60) {
+                        return Err(AgentError::TimeoutWaitingForResponse());
+                    }
+                }
 
                 RequestStatusResponse::Received | RequestStatusResponse::Processing => {
                     if !request_accepted {
@@ -953,10 +952,11 @@ impl Agent {
     }
 
     fn verify_cert_timestamp(&self, cert: &Certificate) -> Result<(), AgentError> {
+        // Verify that the certificate is not older than ingress expiry
+        // Certificates with timestamps in the future are allowed
         let time = lookup_time(cert)?;
         if (OffsetDateTime::now_utc()
             - OffsetDateTime::from_unix_timestamp_nanos(time.into()).unwrap())
-        .abs()
             > self.ingress_expiry
         {
             Err(AgentError::CertificateOutdated(self.ingress_expiry))
@@ -979,12 +979,30 @@ impl Agent {
                     return Err(AgentError::CertificateHasTooManyDelegations);
                 }
                 self.verify_cert(&cert, effective_canister_id)?;
-                let canister_range_lookup = [
-                    "subnet".as_bytes(),
-                    delegation.subnet_id.as_ref(),
-                    "canister_ranges".as_bytes(),
-                ];
-                let canister_range = lookup_value(&cert.tree, canister_range_lookup)?;
+                let canister_range_shards_lookup =
+                    ["canister_ranges".as_bytes(), delegation.subnet_id.as_ref()];
+                let canister_range_shards = lookup_tree(&cert.tree, canister_range_shards_lookup)?;
+                let mut shard_paths = canister_range_shards
+                    .list_paths() // /canister_ranges/<subnet_id>/<shard>
+                    .into_iter()
+                    .map(|mut x| {
+                        x.pop() // flatten [label] to label
+                            .ok_or_else(AgentError::CertificateVerificationFailed)
+                    })
+                    .collect::<Result<Vec<_>, _>>()?;
+                if shard_paths.is_empty() {
+                    return Err(AgentError::CertificateNotAuthorized());
+                }
+                shard_paths.sort_unstable();
+                let shard_division = shard_paths
+                    .partition_point(|shard| shard.as_bytes() <= effective_canister_id.as_slice());
+                if shard_division == 0 {
+                    // the certificate is not authorized to answer calls for this canister
+                    return Err(AgentError::CertificateNotAuthorized());
+                }
+                let max_potential_shard = &shard_paths[shard_division - 1];
+                let canister_range_lookup = [max_potential_shard.as_bytes()];
+                let canister_range = lookup_value(&canister_range_shards, canister_range_lookup)?;
                 let ranges: Vec<(Principal, Principal)> =
                     serde_cbor::from_slice(canister_range).map_err(AgentError::InvalidCborData)?;
                 if !principal_is_within_ranges(&effective_canister_id, &ranges[..]) {
@@ -1018,7 +1036,7 @@ impl Agent {
                 self.verify_cert_for_subnet(&cert, subnet_id)?;
                 let public_key_path = [
                     "subnet".as_bytes(),
-                    delegation.subnet_id.as_ref(),
+                    subnet_id.as_ref(),
                     "public_key".as_bytes(),
                 ];
                 let pk = lookup_value(&cert.tree, public_key_path)
@@ -1101,6 +1119,20 @@ impl Agent {
         lookup_subnet_metrics(cert, subnet_id)
     }
 
+    /// Request a list of metrics about the subnet.
+    pub async fn read_state_subnet_canister_ranges(
+        &self,
+        subnet_id: Principal,
+    ) -> Result<Vec<(Principal, Principal)>, AgentError> {
+        let paths = vec![vec![
+            "subnet".into(),
+            Label::from_bytes(subnet_id.as_slice()),
+            "canister_ranges".into(),
+        ]];
+        let cert = self.read_subnet_state_raw(paths, subnet_id).await?;
+        lookup_subnet_canister_ranges(&cert, subnet_id)
+    }
+
     /// Fetches the status of a particular request by its ID.
     pub async fn request_status_raw(
         &self,
@@ -1142,7 +1174,7 @@ impl Agent {
         &self,
         canister_id: &Principal,
         method_name: S,
-    ) -> UpdateBuilder {
+    ) -> UpdateBuilder<'_> {
         UpdateBuilder::new(self, *canister_id, method_name.into())
     }
 
@@ -1159,7 +1191,11 @@ impl Agent {
 
     /// Returns a `QueryBuilder` enabling the construction of a query call without
     /// passing all arguments.
-    pub fn query<S: Into<String>>(&self, canister_id: &Principal, method_name: S) -> QueryBuilder {
+    pub fn query<S: Into<String>>(
+        &self,
+        canister_id: &Principal,
+        method_name: S,
+    ) -> QueryBuilder<'_> {
         QueryBuilder::new(self, *canister_id, method_name.into())
     }
 
@@ -1185,7 +1221,9 @@ impl Agent {
         })
     }
 
-    async fn get_subnet_by_canister(
+    /// Retrieve subnet information for a canister. This uses an internal five-minute cache, fresh data can
+    /// be fetched with [`fetch_subnet_by_canister`](Self::fetch_subnet_by_canister).
+    pub async fn get_subnet_by_canister(
         &self,
         canister: &Principal,
     ) -> Result<Arc<Subnet>, AgentError> {
@@ -1201,7 +1239,22 @@ impl Agent {
         }
     }
 
-    /// Retrieve all existing API boundary nodes from the state tree via endpoint `/api/v2/canister/<effective_canister_id>/read_state`
+    /// Retrieve subnet information for a subnet ID. This uses an internal five-minute cache, fresh data can
+    /// be fetched with [`fetch_subnet_by_id`](Self::fetch_subnet_by_id).
+    pub async fn get_subnet_by_id(&self, subnet_id: &Principal) -> Result<Arc<Subnet>, AgentError> {
+        let subnet = self
+            .subnet_key_cache
+            .lock()
+            .unwrap()
+            .get_subnet_by_id(subnet_id);
+        if let Some(subnet) = subnet {
+            Ok(subnet)
+        } else {
+            self.fetch_subnet_by_id(subnet_id).await
+        }
+    }
+
+    /// Retrieve all existing API boundary nodes from the state tree via endpoint `/api/v3/canister/<effective_canister_id>/read_state`
     pub async fn fetch_api_boundary_nodes_by_canister_id(
         &self,
         canister_id: Principal,
@@ -1212,7 +1265,7 @@ impl Agent {
         Ok(api_boundary_nodes)
     }
 
-    /// Retrieve all existing API boundary nodes from the state tree via endpoint `/api/v2/subnet/<subnet_id>/read_state`
+    /// Retrieve all existing API boundary nodes from the state tree via endpoint `/api/v3/subnet/<subnet_id>/read_state`
     pub async fn fetch_api_boundary_nodes_by_subnet_id(
         &self,
         subnet_id: Principal,
@@ -1223,20 +1276,66 @@ impl Agent {
         Ok(api_boundary_nodes)
     }
 
-    async fn fetch_subnet_by_canister(
+    /// Fetches and caches the subnet information for a canister.
+    ///
+    /// This function does not read from the cache; most users want
+    /// [`get_subnet_by_canister`](Self::get_subnet_by_canister) instead.
+    pub async fn fetch_subnet_by_canister(
         &self,
         canister: &Principal,
     ) -> Result<Arc<Subnet>, AgentError> {
-        let cert = self
+        let canister_cert = self
             .read_state_raw(vec![vec!["subnet".into()]], *canister)
             .await?;
-
-        let (subnet_id, subnet) = lookup_subnet(&cert, &self.root_key.read().unwrap())?;
+        let subnet_id = if let Some(delegation) = canister_cert.delegation.as_ref() {
+            Principal::from_slice(&delegation.subnet_id)
+        } else {
+            // if no delegation, it comes from the root subnet
+            Principal::self_authenticating(&self.root_key.read().unwrap()[..])
+        };
+        let mut subnet = lookup_incomplete_subnet(&subnet_id, &canister_cert)?;
+        let canister_ranges = if let Some(delegation) = canister_cert.delegation.as_ref() {
+            // non-root subnets will not serve /subnet/<>/canister_ranges when looked up by canister, but their delegation will contain /canister_ranges
+            let delegation_cert: Certificate = serde_cbor::from_slice(&delegation.certificate)?;
+            lookup_canister_ranges(&subnet_id, &delegation_cert)?
+        } else {
+            lookup_canister_ranges(&subnet_id, &canister_cert)?
+        };
+        subnet.canister_ranges = canister_ranges;
+        if !subnet.canister_ranges.contains(canister) {
+            return Err(AgentError::CertificateNotAuthorized());
+        }
         let subnet = Arc::new(subnet);
         self.subnet_key_cache
             .lock()
             .unwrap()
             .insert_subnet(subnet_id, subnet.clone());
+        Ok(subnet)
+    }
+
+    /// Fetches and caches the subnet information for a subnet ID.
+    ///
+    /// This function does not read from the cache; most users want
+    /// [`get_subnet_by_id`](Self::get_subnet_by_id) instead.
+    pub async fn fetch_subnet_by_id(
+        &self,
+        subnet_id: &Principal,
+    ) -> Result<Arc<Subnet>, AgentError> {
+        let subnet_cert = self
+            .read_subnet_state_raw(
+                vec![
+                    vec!["canister_ranges".into(), subnet_id.as_slice().into()],
+                    vec!["subnet".into(), subnet_id.as_slice().into()],
+                ],
+                *subnet_id,
+            )
+            .await?;
+        let subnet = lookup_subnet_and_ranges(subnet_id, &subnet_cert)?;
+        let subnet = Arc::new(subnet);
+        self.subnet_key_cache
+            .lock()
+            .unwrap()
+            .insert_subnet(*subnet_id, subnet.clone());
         Ok(subnet)
     }
 
@@ -1246,14 +1345,25 @@ impl Agent {
         endpoint: &str,
         body: Option<Vec<u8>>,
     ) -> Result<(StatusCode, HeaderMap, Vec<u8>), AgentError> {
-        let create_request_with_generated_url = || -> Result<Request, AgentError> {
+        let body = body.map(Bytes::from);
+
+        let create_request_with_generated_url = || -> Result<http::Request<Bytes>, AgentError> {
             let url = self.route_provider.route()?.join(endpoint)?;
-            let mut http_request = Request::new(method.clone(), url);
-            http_request
-                .headers_mut()
-                .insert(CONTENT_TYPE, "application/cbor".parse().unwrap());
-            *http_request.body_mut() = body.clone().map(Body::from);
-            Ok(http_request)
+            let uri = Uri::from_str(url.as_str())
+                .map_err(|e| AgentError::InvalidReplicaUrl(e.to_string()))?;
+            let body = body.clone().unwrap_or_default();
+            let request = http::Request::builder()
+                .method(method.clone())
+                .uri(uri)
+                .header(CONTENT_TYPE, "application/cbor")
+                .body(body)
+                .map_err(|e| {
+                    AgentError::TransportError(TransportError::Generic(format!(
+                        "unable to create request: {e:#}"
+                    )))
+                })?;
+
+            Ok(request)
         };
 
         let response = self
@@ -1261,40 +1371,13 @@ impl Agent {
             .call(
                 &create_request_with_generated_url,
                 self.max_tcp_error_retries,
+                self.max_response_body_size,
             )
             .await?;
 
-        let http_status = response.status();
-        let response_headers = response.headers().clone();
+        let (parts, body) = response.into_parts();
 
-        // Size Check (Content-Length)
-        if matches!(self
-            .max_response_body_size
-            .zip(response.content_length()), Some((size_limit, content_length)) if content_length > size_limit as u64)
-        {
-            return Err(AgentError::ResponseSizeExceededLimit());
-        }
-
-        let mut body: Vec<u8> = response
-            .content_length()
-            .map_or_else(Vec::new, |n| Vec::with_capacity(n as usize));
-
-        let mut stream = response.bytes_stream();
-
-        while let Some(chunk) = stream.next().await {
-            let chunk = chunk?;
-
-            // Size Check (Body Size)
-            if matches!(self
-                .max_response_body_size, Some(size_limit) if body.len() + chunk.len() > size_limit)
-            {
-                return Err(AgentError::ResponseSizeExceededLimit());
-            }
-
-            body.extend_from_slice(chunk.as_ref());
-        }
-
-        Ok((http_status, response_headers, body))
+        Ok((parts.status, parts.headers, body.to_vec()))
     }
 
     async fn execute(
@@ -1574,7 +1657,7 @@ struct SubnetCache {
 impl SubnetCache {
     fn new() -> Self {
         Self {
-            subnets: TimedCache::with_lifespan(300),
+            subnets: TimedCache::with_lifespan(Duration::from_secs(300)),
             canister_index: RangeInclusiveMap::new_with_step_fns(),
         }
     }
@@ -1586,6 +1669,10 @@ impl SubnetCache {
             .filter(|subnet| subnet.canister_ranges.contains(canister))
     }
 
+    fn get_subnet_by_id(&mut self, subnet_id: &Principal) -> Option<Arc<Subnet>> {
+        self.subnets.cache_get(subnet_id).cloned()
+    }
+
     fn insert_subnet(&mut self, subnet_id: Principal, subnet: Arc<Subnet>) {
         self.subnets.cache_set(subnet_id, subnet.clone());
         for range in subnet.canister_ranges.iter() {
@@ -1595,7 +1682,7 @@ impl SubnetCache {
 }
 
 #[derive(Clone, Copy)]
-struct PrincipalStep;
+pub(crate) struct PrincipalStep;
 
 impl StepFns<Principal> for PrincipalStep {
     fn add_one(start: &Principal) -> Principal {
@@ -1622,15 +1709,6 @@ impl StepFns<Principal> for PrincipalStep {
         }
         Principal::from_slice(&arr[..bytes.len()])
     }
-}
-
-#[derive(Clone)]
-pub(crate) struct Subnet {
-    // This key is just fetched for completeness. Do not actually use this value as it is not authoritative in case of a rogue subnet.
-    // If a future agent needs to know the subnet key then it should fetch /subnet from the *root* subnet.
-    _key: Vec<u8>,
-    node_keys: HashMap<Principal, Vec<u8>>,
-    canister_ranges: RangeInclusiveSet<Principal, PrincipalStep>,
 }
 
 /// API boundary node, which routes /api calls to IC replica nodes.
@@ -2009,13 +2087,89 @@ impl<'agent> IntoFuture for UpdateBuilder<'agent> {
 #[cfg_attr(target_family = "wasm", async_trait(?Send))]
 #[cfg_attr(not(target_family = "wasm"), async_trait)]
 pub trait HttpService: Send + Sync + Debug {
-    /// Perform a HTTP request. Any retry logic should call `req` again, instead of `Request::try_clone`.
+    /// Perform a HTTP request. Any retry logic should call `req` again to get a new request.
     async fn call<'a>(
         &'a self,
-        req: &'a (dyn Fn() -> Result<Request, AgentError> + Send + Sync),
+        req: &'a (dyn Fn() -> Result<http::Request<Bytes>, AgentError> + Send + Sync),
         max_retries: usize,
-    ) -> Result<Response, AgentError>;
+        size_limit: Option<usize>,
+    ) -> Result<http::Response<Bytes>, AgentError>;
 }
+
+/// Convert from http Request to reqwest's one
+fn from_http_request(req: http::Request<Bytes>) -> Result<Request, AgentError> {
+    let (parts, body) = req.into_parts();
+    let body = reqwest::Body::from(body);
+    // I think it can never fail since it converts from `Url` to `Uri` and `Url` is a subset of `Uri`,
+    // but just to be safe let's handle it.
+    let request = http::Request::from_parts(parts, body)
+        .try_into()
+        .map_err(|e: reqwest::Error| AgentError::InvalidReplicaUrl(e.to_string()))?;
+
+    Ok(request)
+}
+
+/// Convert from reqwests's Response to http one
+#[cfg(not(target_family = "wasm"))]
+async fn to_http_response(
+    resp: Response,
+    size_limit: Option<usize>,
+) -> Result<http::Response<Bytes>, AgentError> {
+    use http_body_util::{BodyExt, Limited};
+
+    let resp: http::Response<reqwest::Body> = resp.into();
+    let (parts, body) = resp.into_parts();
+    let body = Limited::new(body, size_limit.unwrap_or(usize::MAX));
+    let body = body
+        .collect()
+        .await
+        .map_err(|e| {
+            AgentError::TransportError(TransportError::Generic(format!(
+                "unable to read response body: {e:#}"
+            )))
+        })?
+        .to_bytes();
+    let resp = http::Response::from_parts(parts, body);
+
+    Ok(resp)
+}
+
+/// Convert from reqwests's Response to http one.
+/// WASM Response in reqwest doesn't have direct conversion for http::Response,
+/// so we have to hack around using streams.
+#[cfg(target_family = "wasm")]
+async fn to_http_response(
+    resp: Response,
+    size_limit: Option<usize>,
+) -> Result<http::Response<Bytes>, AgentError> {
+    use futures_util::StreamExt;
+    use http_body::Frame;
+    use http_body_util::{Limited, StreamBody};
+
+    // Save headers
+    let status = resp.status();
+    let headers = resp.headers().clone();
+
+    // Convert body
+    let stream = resp.bytes_stream().map(|x| x.map(Frame::data));
+    let body = StreamBody::new(stream);
+    let body = Limited::new(body, size_limit.unwrap_or(usize::MAX));
+    let body = http_body_util::BodyExt::collect(body)
+        .await
+        .map_err(|e| {
+            AgentError::TransportError(TransportError::Generic(format!(
+                "unable to read response body: {e:#}"
+            )))
+        })?
+        .to_bytes();
+
+    let mut resp = http::Response::new(body);
+    *resp.status_mut() = status;
+    *resp.headers_mut() = headers;
+
+    Ok(resp)
+}
+
 #[cfg(not(target_family = "wasm"))]
 #[async_trait]
 impl<T> HttpService for T
@@ -2027,22 +2181,33 @@ where
     #[allow(clippy::needless_arbitrary_self_type)]
     async fn call<'a>(
         mut self: &'a Self,
-        req: &'a (dyn Fn() -> Result<Request, AgentError> + Send + Sync),
+        req: &'a (dyn Fn() -> Result<http::Request<Bytes>, AgentError> + Send + Sync),
         max_retries: usize,
-    ) -> Result<Response, AgentError> {
+        size_limit: Option<usize>,
+    ) -> Result<http::Response<Bytes>, AgentError> {
         let mut retry_count = 0;
         loop {
-            match Service::call(&mut self, req()?).await {
+            let request = from_http_request(req()?)?;
+
+            match Service::call(&mut self, request).await {
                 Err(err) => {
                     // Network-related errors can be retried.
                     if err.is_connect() {
                         if retry_count >= max_retries {
-                            return Err(AgentError::TransportError(err));
+                            return Err(AgentError::TransportError(TransportError::Reqwest(err)));
                         }
                         retry_count += 1;
                     }
+                    // All other errors return immediately.
+                    else {
+                        return Err(AgentError::TransportError(TransportError::Reqwest(err)));
+                    }
                 }
-                Ok(resp) => return Ok(resp),
+
+                Ok(resp) => {
+                    let resp = to_http_response(resp, size_limit).await?;
+                    return Ok(resp);
+                }
             }
         }
     }
@@ -2058,10 +2223,16 @@ where
     #[allow(clippy::needless_arbitrary_self_type)]
     async fn call<'a>(
         mut self: &'a Self,
-        req: &'a (dyn Fn() -> Result<Request, AgentError> + Send + Sync),
-        _: usize,
-    ) -> Result<Response, AgentError> {
-        Ok(Service::call(&mut self, req()?).await?)
+        req: &'a (dyn Fn() -> Result<http::Request<Bytes>, AgentError> + Send + Sync),
+        _retries: usize,
+        _size_limit: Option<usize>,
+    ) -> Result<http::Response<Bytes>, AgentError> {
+        let request = from_http_request(req()?)?;
+        let response = Service::call(&mut self, request)
+            .await
+            .map_err(|e| AgentError::TransportError(TransportError::Reqwest(e)))?;
+
+        to_http_response(response, _size_limit).await
     }
 }
 
@@ -2075,16 +2246,27 @@ struct Retry429Logic {
 impl HttpService for Retry429Logic {
     async fn call<'a>(
         &'a self,
-        req: &'a (dyn Fn() -> Result<Request, AgentError> + Send + Sync),
+        req: &'a (dyn Fn() -> Result<http::Request<Bytes>, AgentError> + Send + Sync),
         _max_tcp_retries: usize,
-    ) -> Result<Response, AgentError> {
+        _size_limit: Option<usize>,
+    ) -> Result<http::Response<Bytes>, AgentError> {
         let mut retries = 0;
         loop {
             #[cfg(not(target_family = "wasm"))]
-            let resp = self.client.call(req, _max_tcp_retries).await?;
+            let resp = self.client.call(req, _max_tcp_retries, _size_limit).await?;
             // Client inconveniently does not implement Service on wasm
             #[cfg(target_family = "wasm")]
-            let resp = self.client.execute(req()?).await?;
+            let resp = {
+                let request = from_http_request(req()?)?;
+                let resp = self
+                    .client
+                    .execute(request)
+                    .await
+                    .map_err(|e| AgentError::TransportError(TransportError::Reqwest(e)))?;
+
+                to_http_response(resp, _size_limit).await?
+            };
+
             if resp.status() == StatusCode::TOO_MANY_REQUESTS {
                 if retries == 6 {
                     break Ok(resp);
