@@ -23,11 +23,8 @@ pub(super) fn route_n_times(n: usize, f: Arc<impl RouteProvider + ?Sized>) -> Ve
         .collect()
 }
 
-pub(super) fn assert_routed_domains<T>(
-    actual: Vec<T>,
-    expected: Vec<&str>,
-    expected_repetitions: usize,
-) where
+pub(super) fn assert_routed_domains<T>(actual: Vec<T>, expected: Vec<&str>)
+where
     T: AsRef<str> + Eq + Hash + Debug + Ord,
 {
     fn build_count_map<T>(items: &[T]) -> HashMap<&str, usize>
@@ -49,11 +46,15 @@ pub(super) fn assert_routed_domains<T>(
     // Assert all routed domains are present.
     assert_eq!(keys_actual, keys_expected);
 
-    // Assert the expected repetition count of each routed domain.
-    let actual_repetitions = count_actual.values().collect::<Vec<_>>();
-    assert!(actual_repetitions
-        .iter()
-        .all(|&x| x == &expected_repetitions));
+    // For latency-based routing, we can't expect exact equal distribution,
+    // so we just verify that all expected nodes were used at least once.
+    // The probabilistic nature of latency-based routing means distribution will vary.
+    for expected_node in expected {
+        assert!(
+            count_actual.contains_key(expected_node),
+            "Expected node '{expected_node}' was not routed to"
+        );
+    }
 }
 
 #[derive(Debug)]
