@@ -23,9 +23,9 @@ pub enum PemError {
     #[error("An error occurred while reading the file: {0}")]
     DerError(#[from] der::Error),
 
-    /// The key was rejected by ed25519-consensus.
-    #[error("A key was rejected by ed25519-consensus: {0}")]
-    KeyRejected(#[from] ed25519_consensus::Error),
+    /// The private key is invalid.
+    #[error("Invalid private key: {0}")]
+    InvalidPrivateKey(String),
 
     /// The key was rejected by k256.
     #[error("A key was rejected by k256: {0}")]
