@@ -18,68 +18,53 @@
 //!
 //! ## Example
 //! The following example illustrates how to use the Agent interface to send
-//! a call to an Internet Computer canister that performs network management
-//! operations. In this example, the call to the Internet Computer management
-//! canister (`aaaaa-aa`) creates a placeholder for a new canister by
-//! registering a network-specific identifier. The management canister then
-//! returns the result in the form of the textual representation of the canister
-//! identifier to the caller.
+//! a call to an Internet Computer's Ledger Canister to check the total ICP tokens supply.
 //!
-//! ```ignore
-//! # // This test is ignored because it requires an ic to be running. We run these
-//! # // in the ic-ref workflow.
-//! use ic_agent::{Agent, export::Principal};
-//! use candid::{Encode, Decode, CandidType, Nat};
-//! use serde::Deserialize;
+//! ```rust
+//!use anyhow::{Context, Result};
+//!use candid::{Decode, Nat};
+//!use ic_agent::{export::Principal, Agent};
+//!use url::Url;
 //!
-//! #[derive(CandidType)]
-//! struct Argument {
-//!   amount: Option<Nat>,
-//! }
+//!pub async fn create_agent(url: Url, use_mainnet: bool) -> Result<Agent> {
+//!    let agent = Agent::builder().with_url(url).build()?;
+//!    if !use_mainnet {
+//!        agent.fetch_root_key().await?;
+//!    }
+//!    Ok(agent)
+//!}
 //!
-//! #[derive(CandidType, Deserialize)]
-//! struct CreateCanisterResult {
-//!   canister_id: Principal,
-//! }
+//!#[tokio::main]
+//!async fn main() -> Result<()> {
+//!    // IC HTTP Gateway URL
+//!    let url = Url::parse("https://ic0.app").unwrap();
+//!    let agent = create_agent(url, true).await?;
 //!
-//! # fn create_identity() -> impl ic_agent::Identity {
-//! #     ic_agent::identity::BasicIdentity::from_signing_key(
-//! #         ed25519_consensus::SigningKey::new(rand::thread_rng()),
-//! #     )
-//! # }
-//! #
-//! async fn create_a_canister() -> Result<Principal, Box<dyn std::error::Error>> {
-//! # let url = format!("http://localhost:{}", option_env!("IC_REF_PORT").unwrap_or("4943"));
-//!   let agent = Agent::builder()
-//!     .with_url(url)
-//!     .with_identity(create_identity())
-//!     .build()?;
-//!   // Only do the following call when not contacting the IC main net (e.g. a local emulator).
-//!   // This is important as the main net public key is static and a rogue network could return
-//!   // a different key.
-//!   // If you know the root key ahead of time, you can use `agent.set_root_key(root_key);`.
-//!   agent.fetch_root_key().await?;
-//!   let management_canister_id = Principal::from_text("aaaaa-aa")?;
+//!    // ICP Ledger Canister ID
+//!    let canister_id = Principal::from_text("ryjl3-tyaaa-aaaaa-aaaba-cai")?;
 //!
-//!   // Create a call to the management canister to create a new canister ID,
-//!   // and wait for a result.
-//!   // The effective canister id must belong to the canister ranges of the subnet at which the canister is created.
-//!   let effective_canister_id = Principal::from_text("rwlgt-iiaaa-aaaaa-aaaaa-cai").unwrap();
-//!   let response = agent.update(&management_canister_id, "provisional_create_canister_with_cycles")
-//!     .with_effective_canister_id(effective_canister_id)
-//!     .with_arg(Encode!(&Argument { amount: None})?)
-//!     .await?;
+//!    // Method: icrc1_total_supply (takes no arguments, returns nat)
+//!    let method_name = "icrc1_total_supply";
 //!
-//!   let result = Decode!(response.as_slice(), CreateCanisterResult)?;
-//!   let canister_id: Principal = result.canister_id;
-//!   Ok(canister_id)
-//! }
+//!    // Encode empty Candid arguments
+//!    let args = candid::encode_args(())?;
 //!
-//! # let mut runtime = tokio::runtime::Runtime::new().unwrap();
-//! # runtime.block_on(async {
-//! let canister_id = create_a_canister().await.unwrap();
-//! eprintln!("{}", canister_id);
-//! # });
+//!    // Dispatch query call
+//!    let response = agent
+//!        .query(&canister_id, method_name)
+//!        .with_arg(args)
+//!        .call()
+//!        .await
+//!        .context("Failed to query icrc1_total_supply method.")?;
+//!
+//!    // Decode the response as nat
+//!    let total_supply_nat =
+//!        Decode!(&response, Nat).context("Failed to decode total supply as nat.")?;
+//!
+//!    println!("Total ICP Supply: {} ICP", total_supply_nat);
+//!
+//!    Ok(())
+//!}
 //! ```
 //! For more information about the Agent interface used in this example, see the
 //! [Agent] documentation.
@@ -102,7 +87,7 @@
     rustdoc::private_intra_doc_links
 )]
 #![cfg_attr(not(target_family = "wasm"), warn(clippy::future_not_send))]
-#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 #[macro_use]
 mod util;
