@@ -6,8 +6,21 @@ use crate::{
     call::{AsyncCall, SyncCall},
     Canister,
 };
-use candid::{CandidType, Deserialize, Nat};
 use ic_agent::{export::Principal, Agent};
+use ic_management_canister_types::{
+    CanisterIdRecord, DeleteCanisterSnapshotArgs, LoadCanisterSnapshotArgs,
+    ProvisionalTopUpCanisterArgs, ReadCanisterSnapshotDataArgs, ReadCanisterSnapshotMetadataArgs,
+    TakeCanisterSnapshotArgs, UploadCanisterSnapshotDataArgs, UploadCanisterSnapshotMetadataArgs,
+    UploadChunkArgs,
+};
+// Re-export the types that are used be defined in this file.
+pub use ic_management_canister_types::{
+    CanisterLogRecord, CanisterStatusResult, CanisterStatusType, CanisterTimer, ChunkHash,
+    DefiniteCanisterSettings, FetchCanisterLogsResult, LogVisibility, OnLowWasmMemoryHookStatus,
+    QueryStats, ReadCanisterSnapshotDataResult, ReadCanisterSnapshotMetadataResult, Snapshot,
+    SnapshotDataKind, SnapshotDataOffset, SnapshotMetadataGlobal, SnapshotSource,
+    StoredChunksResult, UploadCanisterSnapshotMetadataResult, UploadChunkResult,
+};
 use std::{convert::AsRef, ops::Deref};
 use strum_macros::{AsRefStr, Display, EnumString};
 
@@ -77,6 +90,14 @@ pub enum MgmtMethod {
     ListCanisterSnapshots,
     /// See [`ManagementCanister::delete_canister_snapshot`].
     DeleteCanisterSnapshot,
+    /// See [`ManagementCanister::read_canister_snapshot_metadata`].
+    ReadCanisterSnapshotMetadata,
+    /// See [`ManagementCanister::read_canister_snapshot_data`].
+    ReadCanisterSnapshotData,
+    /// See [`ManagementCanister::upload_canister_snapshot_metadata`].
+    UploadCanisterSnapshotMetadata,
+    /// See [`ManagementCanister::upload_canister_snapshot_data`].
+    UploadCanisterSnapshotData,
     /// There is no corresponding agent function as only canisters can call it.
     EcdsaPublicKey,
     /// There is no corresponding agent function as only canisters can call it.
@@ -115,171 +136,56 @@ impl<'agent> ManagementCanister<'agent> {
     }
 }
 
-/// The complete canister status information of a canister. This includes
-/// the `CanisterStatus`, a hash of the module installed on the canister (None if nothing installed),
-/// the controller of the canister, the canister's memory size, and its balance in cycles.
-#[derive(Clone, Debug, Deserialize, CandidType)]
-pub struct StatusCallResult {
-    /// The status of the canister.
-    pub status: CanisterStatus,
-    /// The canister's settings.
-    pub settings: DefiniteCanisterSettings,
-    /// The SHA-256 hash of the canister's installed code, if any.
-    pub module_hash: Option<Vec<u8>>,
-    /// The total size, in bytes, of the memory the canister is using.
-    pub memory_size: Nat,
-    /// The canister's cycle balance.
-    pub cycles: Nat,
-    /// The canister's reserved cycles balance.
-    pub reserved_cycles: Nat,
-    /// The cycles burned by the canister in one day for its resource usage
-    /// (compute and memory allocation and memory usage).
-    pub idle_cycles_burned_per_day: Nat,
-    /// Additional information relating to query calls.
-    pub query_stats: QueryStats,
-}
+#[doc(hidden)]
+#[deprecated(since = "0.42.0", note = "Please use CanisterStatusResult instead")]
+pub type StatusCallResult = CanisterStatusResult;
 
-/// Statistics relating to query calls.
-#[derive(Clone, Debug, Deserialize, CandidType)]
-pub struct QueryStats {
-    /// The total number of query calls this canister has performed.
-    pub num_calls_total: Nat,
-    /// The total number of instructions this canister has executed during query calls.
-    pub num_instructions_total: Nat,
-    /// The total number of bytes in request payloads sent to this canister's query calls.
-    pub request_payload_bytes_total: Nat,
-    /// The total number of bytes in response payloads returned from this canister's query calls.
-    pub response_payload_bytes_total: Nat,
-}
+#[doc(hidden)]
+#[deprecated(since = "0.42.0", note = "Please use CanisterStatusType instead")]
+pub type CanisterStatus = CanisterStatusType;
 
-/// Log visibility for a canister.
-#[derive(Default, Clone, CandidType, Deserialize, Debug, PartialEq, Eq)]
-pub enum LogVisibility {
-    #[default]
-    #[serde(rename = "controllers")]
-    /// Canister logs are visible to controllers only.
-    Controllers,
-    #[serde(rename = "public")]
-    /// Canister logs are visible to everyone.
-    Public,
-    #[serde(rename = "allowed_viewers")]
-    /// Canister logs are visible to a set of principals.
-    AllowedViewers(Vec<Principal>),
-}
+#[doc(hidden)]
+#[deprecated(since = "0.42.0", note = "Please use FetchCanisterLogsResult instead")]
+pub type FetchCanisterLogsResponse = FetchCanisterLogsResult;
 
-/// The concrete settings of a canister.
-#[derive(Clone, Debug, Deserialize, CandidType)]
-pub struct DefiniteCanisterSettings {
-    /// The set of canister controllers. Controllers can update the canister via the management canister.
-    pub controllers: Vec<Principal>,
-    /// The allocation percentage (between 0 and 100 inclusive) for *guaranteed* compute capacity.
-    pub compute_allocation: Nat,
-    /// The allocation, in bytes (up to 256 TiB) that the canister is allowed to use for storage.
-    pub memory_allocation: Nat,
-    /// The IC will freeze a canister protectively if it will likely run out of cycles before this amount of time,
-    /// in seconds (up to `u64::MAX`), has passed.
-    pub freezing_threshold: Nat,
-    /// The upper limit of the canister's reserved cycles balance.
-    pub reserved_cycles_limit: Option<Nat>,
-    /// A soft limit on the Wasm memory usage of the canister in bytes (up to 256TiB).
-    pub wasm_memory_limit: Option<Nat>,
-    /// A threshold on the Wasm memory usage of the canister as a distance in bytes from `wasm_memory_limit`,
-    /// at which the canister's `on_low_wasm_memory` hook will be called (up to 256TiB)
-    pub wasm_memory_threshold: Option<Nat>,
-    /// The canister log visibility. Defines which principals are allowed to fetch logs.
-    pub log_visibility: LogVisibility,
-}
+#[doc(hidden)]
+#[deprecated(since = "0.42.0", note = "Please use StoredChunksResult instead")]
+pub type StoreChunksResult = StoredChunksResult;
 
-impl std::fmt::Display for StatusCallResult {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        std::fmt::Debug::fmt(self, f)
-    }
-}
+#[doc(hidden)]
+#[deprecated(
+    since = "0.42.0",
+    note = "Please use ReadCanisterSnapshotMetadataResult instead"
+)]
+pub type SnapshotMetadata = ReadCanisterSnapshotMetadataResult;
 
-/// The status of a Canister, whether it's running, in the process of stopping, or
-/// stopped.
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq, CandidType)]
-pub enum CanisterStatus {
-    /// The canister is currently running.
-    #[serde(rename = "running")]
-    Running,
-    /// The canister is in the process of stopping.
-    #[serde(rename = "stopping")]
-    Stopping,
-    /// The canister is stopped.
-    #[serde(rename = "stopped")]
-    Stopped,
-}
+#[doc(hidden)]
+#[deprecated(
+    since = "0.42.0",
+    note = "Please use ReadCanisterSnapshotDataResult instead"
+)]
+pub type SnapshotDataResult = ReadCanisterSnapshotDataResult;
 
-impl std::fmt::Display for CanisterStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        std::fmt::Debug::fmt(self, f)
-    }
-}
-
-/// A log record of a canister.
-#[derive(Default, Clone, CandidType, Deserialize, Debug, PartialEq, Eq)]
-pub struct CanisterLogRecord {
-    /// The index of the log record.
-    pub idx: u64,
-    /// The timestamp of the log record.
-    pub timestamp_nanos: u64,
-    /// The content of the log record.
-    #[serde(with = "serde_bytes")]
-    pub content: Vec<u8>,
-}
-
-/// The result of a [`ManagementCanister::fetch_canister_logs`] call.
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq, CandidType)]
-pub struct FetchCanisterLogsResponse {
-    /// The logs of the canister.
-    pub canister_log_records: Vec<CanisterLogRecord>,
-}
-
-/// Chunk hash.
-#[derive(CandidType, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone)]
-pub struct ChunkHash {
-    /// The hash of an uploaded chunk
-    #[serde(with = "serde_bytes")]
-    pub hash: Vec<u8>,
-}
-
-/// Return type of [`ManagementCanister::stored_chunks`].
-pub type StoreChunksResult = Vec<ChunkHash>;
-
-/// Return type of [`ManagementCanister::upload_chunk`].
-pub type UploadChunkResult = ChunkHash;
-
-/// A recorded snapshot of a canister. Can be restored with [`ManagementCanister::load_canister_snapshot`].
-#[derive(Debug, Clone, CandidType, Deserialize)]
-pub struct Snapshot {
-    /// The ID of the snapshot.
-    #[serde(with = "serde_bytes")]
-    pub id: Vec<u8>,
-    /// The Unix nanosecond timestamp the snapshot was taken at.
-    pub taken_at_timestamp: u64,
-    /// The size of the snapshot in bytes.
-    pub total_size: u64,
-}
+#[doc(hidden)]
+#[deprecated(
+    since = "0.42.0",
+    note = "Please use UploadCanisterSnapshotMetadataResult instead"
+)]
+pub type CanisterSnapshotId = UploadCanisterSnapshotMetadataResult;
 
 impl<'agent> ManagementCanister<'agent> {
     /// Get the status of a canister.
     pub fn canister_status(
         &self,
         canister_id: &Principal,
-    ) -> impl 'agent + AsyncCall<Value = (StatusCallResult,)> {
-        #[derive(CandidType)]
-        struct In {
-            canister_id: Principal,
-        }
-
+    ) -> impl 'agent + AsyncCall<Value = (CanisterStatusResult,)> {
         self.update(MgmtMethod::CanisterStatus.as_ref())
-            .with_arg(In {
+            .with_arg(CanisterIdRecord {
                 canister_id: *canister_id,
             })
             .with_effective_canister_id(canister_id.to_owned())
             .build()
-            .map(|result: (StatusCallResult,)| (result.0,))
+            .map(|result: (CanisterStatusResult,)| (result.0,))
     }
 
     /// Create a canister.
@@ -290,13 +196,8 @@ impl<'agent> ManagementCanister<'agent> {
     /// This method deposits the cycles included in this call into the specified canister.
     /// Only the controller of the canister can deposit cycles.
     pub fn deposit_cycles(&self, canister_id: &Principal) -> impl 'agent + AsyncCall<Value = ()> {
-        #[derive(CandidType)]
-        struct Argument {
-            canister_id: Principal,
-        }
-
         self.update(MgmtMethod::DepositCycles.as_ref())
-            .with_arg(Argument {
+            .with_arg(CanisterIdRecord {
                 canister_id: *canister_id,
             })
             .with_effective_canister_id(canister_id.to_owned())
@@ -305,13 +206,8 @@ impl<'agent> ManagementCanister<'agent> {
 
     /// Deletes a canister.
     pub fn delete_canister(&self, canister_id: &Principal) -> impl 'agent + AsyncCall<Value = ()> {
-        #[derive(CandidType)]
-        struct Argument {
-            canister_id: Principal,
-        }
-
         self.update(MgmtMethod::DeleteCanister.as_ref())
-            .with_arg(Argument {
+            .with_arg(CanisterIdRecord {
                 canister_id: *canister_id,
             })
             .with_effective_canister_id(canister_id.to_owned())
@@ -325,19 +221,10 @@ impl<'agent> ManagementCanister<'agent> {
     pub fn provisional_top_up_canister(
         &self,
         canister_id: &Principal,
-        amount: u64,
+        top_up_args: &ProvisionalTopUpCanisterArgs,
     ) -> impl 'agent + AsyncCall<Value = ()> {
-        #[derive(CandidType)]
-        struct Argument {
-            canister_id: Principal,
-            amount: u64,
-        }
-
         self.update(MgmtMethod::ProvisionalTopUpCanister.as_ref())
-            .with_arg(Argument {
-                canister_id: *canister_id,
-                amount,
-            })
+            .with_arg(top_up_args)
             .with_effective_canister_id(canister_id.to_owned())
             .build()
     }
@@ -353,13 +240,8 @@ impl<'agent> ManagementCanister<'agent> {
 
     /// Starts a canister.
     pub fn start_canister(&self, canister_id: &Principal) -> impl 'agent + AsyncCall<Value = ()> {
-        #[derive(CandidType)]
-        struct Argument {
-            canister_id: Principal,
-        }
-
         self.update(MgmtMethod::StartCanister.as_ref())
-            .with_arg(Argument {
+            .with_arg(CanisterIdRecord {
                 canister_id: *canister_id,
             })
             .with_effective_canister_id(canister_id.to_owned())
@@ -368,13 +250,8 @@ impl<'agent> ManagementCanister<'agent> {
 
     /// Stop a canister.
     pub fn stop_canister(&self, canister_id: &Principal) -> impl 'agent + AsyncCall<Value = ()> {
-        #[derive(CandidType)]
-        struct Argument {
-            canister_id: Principal,
-        }
-
         self.update(MgmtMethod::StopCanister.as_ref())
-            .with_arg(Argument {
+            .with_arg(CanisterIdRecord {
                 canister_id: *canister_id,
             })
             .with_effective_canister_id(canister_id.to_owned())
@@ -389,13 +266,8 @@ impl<'agent> ManagementCanister<'agent> {
     /// The canister is now empty. In particular, any incoming or queued calls will be rejected.
     //// A canister after uninstalling retains its cycles balance, controller, status, and allocations.
     pub fn uninstall_code(&self, canister_id: &Principal) -> impl 'agent + AsyncCall<Value = ()> {
-        #[derive(CandidType)]
-        struct Argument {
-            canister_id: Principal,
-        }
-
         self.update(MgmtMethod::UninstallCode.as_ref())
-            .with_arg(Argument {
+            .with_arg(CanisterIdRecord {
                 canister_id: *canister_id,
             })
             .with_effective_canister_id(canister_id.to_owned())
@@ -423,20 +295,10 @@ impl<'agent> ManagementCanister<'agent> {
     pub fn upload_chunk(
         &self,
         canister_id: &Principal,
-        chunk: &[u8],
+        upload_chunk_args: &UploadChunkArgs,
     ) -> impl 'agent + AsyncCall<Value = (UploadChunkResult,)> {
-        #[derive(CandidType, Deserialize)]
-        struct Argument<'a> {
-            canister_id: Principal,
-            #[serde(with = "serde_bytes")]
-            chunk: &'a [u8],
-        }
-
         self.update(MgmtMethod::UploadChunk.as_ref())
-            .with_arg(Argument {
-                canister_id: *canister_id,
-                chunk,
-            })
+            .with_arg(upload_chunk_args)
             .with_effective_canister_id(*canister_id)
             .build()
     }
@@ -446,12 +308,10 @@ impl<'agent> ManagementCanister<'agent> {
         &self,
         canister_id: &Principal,
     ) -> impl 'agent + AsyncCall<Value = ()> {
-        #[derive(CandidType)]
-        struct Argument<'a> {
-            canister_id: &'a Principal,
-        }
         self.update(MgmtMethod::ClearChunkStore.as_ref())
-            .with_arg(Argument { canister_id })
+            .with_arg(CanisterIdRecord {
+                canister_id: *canister_id,
+            })
             .with_effective_canister_id(*canister_id)
             .build()
     }
@@ -460,13 +320,11 @@ impl<'agent> ManagementCanister<'agent> {
     pub fn stored_chunks(
         &self,
         canister_id: &Principal,
-    ) -> impl 'agent + AsyncCall<Value = (StoreChunksResult,)> {
-        #[derive(CandidType)]
-        struct Argument<'a> {
-            canister_id: &'a Principal,
-        }
+    ) -> impl 'agent + AsyncCall<Value = (StoredChunksResult,)> {
         self.update(MgmtMethod::StoredChunks.as_ref())
-            .with_arg(Argument { canister_id })
+            .with_arg(CanisterIdRecord {
+                canister_id: *canister_id,
+            })
             .with_effective_canister_id(*canister_id)
             .build()
     }
@@ -497,15 +355,10 @@ impl<'agent> ManagementCanister<'agent> {
     pub fn fetch_canister_logs(
         &self,
         canister_id: &Principal,
-    ) -> impl 'agent + SyncCall<Value = (FetchCanisterLogsResponse,)> {
-        #[derive(CandidType)]
-        struct In {
-            canister_id: Principal,
-        }
-
+    ) -> impl 'agent + SyncCall<Value = (FetchCanisterLogsResult,)> {
         // `fetch_canister_logs` is only supported in non-replicated mode.
         self.query(MgmtMethod::FetchCanisterLogs.as_ref())
-            .with_arg(In {
+            .with_arg(CanisterIdRecord {
                 canister_id: *canister_id,
             })
             .with_effective_canister_id(*canister_id)
@@ -518,18 +371,10 @@ impl<'agent> ManagementCanister<'agent> {
     pub fn take_canister_snapshot(
         &self,
         canister_id: &Principal,
-        replace_snapshot: Option<&[u8]>,
+        take_args: &TakeCanisterSnapshotArgs,
     ) -> impl 'agent + AsyncCall<Value = (Snapshot,)> {
-        #[derive(CandidType)]
-        struct In<'a> {
-            canister_id: Principal,
-            replace_snapshot: Option<&'a [u8]>,
-        }
         self.update(MgmtMethod::TakeCanisterSnapshot.as_ref())
-            .with_arg(In {
-                canister_id: *canister_id,
-                replace_snapshot,
-            })
+            .with_arg(take_args)
             .with_effective_canister_id(*canister_id)
             .build()
     }
@@ -540,20 +385,10 @@ impl<'agent> ManagementCanister<'agent> {
     pub fn load_canister_snapshot(
         &self,
         canister_id: &Principal,
-        snapshot_id: &[u8],
+        load_args: &LoadCanisterSnapshotArgs,
     ) -> impl 'agent + AsyncCall<Value = ()> {
-        #[derive(CandidType)]
-        struct In<'a> {
-            canister_id: Principal,
-            snapshot_id: &'a [u8],
-            sender_canister_version: Option<u64>,
-        }
         self.update(MgmtMethod::LoadCanisterSnapshot.as_ref())
-            .with_arg(In {
-                canister_id: *canister_id,
-                snapshot_id,
-                sender_canister_version: None,
-            })
+            .with_arg(load_args)
             .with_effective_canister_id(*canister_id)
             .build()
     }
@@ -563,12 +398,8 @@ impl<'agent> ManagementCanister<'agent> {
         &self,
         canister_id: &Principal,
     ) -> impl 'agent + AsyncCall<Value = (Vec<Snapshot>,)> {
-        #[derive(CandidType)]
-        struct In {
-            canister_id: Principal,
-        }
         self.update(MgmtMethod::ListCanisterSnapshots.as_ref())
-            .with_arg(In {
+            .with_arg(CanisterIdRecord {
                 canister_id: *canister_id,
             })
             .with_effective_canister_id(*canister_id)
@@ -579,18 +410,59 @@ impl<'agent> ManagementCanister<'agent> {
     pub fn delete_canister_snapshot(
         &self,
         canister_id: &Principal,
-        snapshot_id: &[u8],
+        delete_args: &DeleteCanisterSnapshotArgs,
     ) -> impl 'agent + AsyncCall<Value = ()> {
-        #[derive(CandidType)]
-        struct In<'a> {
-            canister_id: Principal,
-            snapshot_id: &'a [u8],
-        }
         self.update(MgmtMethod::DeleteCanisterSnapshot.as_ref())
-            .with_arg(In {
-                canister_id: *canister_id,
-                snapshot_id,
-            })
+            .with_arg(delete_args)
+            .with_effective_canister_id(*canister_id)
+            .build()
+    }
+
+    /// Reads the metadata of a recorded canister snapshot by canister ID and snapshot ID.
+    pub fn read_canister_snapshot_metadata(
+        &self,
+        canister_id: &Principal,
+        metadata_args: &ReadCanisterSnapshotMetadataArgs,
+    ) -> impl 'agent + AsyncCall<Value = (ReadCanisterSnapshotMetadataResult,)> {
+        self.update(MgmtMethod::ReadCanisterSnapshotMetadata.as_ref())
+            .with_arg(metadata_args)
+            .with_effective_canister_id(*canister_id)
+            .build()
+    }
+
+    /// Reads the data of a recorded canister snapshot by canister ID and snapshot ID.
+    pub fn read_canister_snapshot_data(
+        &self,
+        canister_id: &Principal,
+        data_args: &ReadCanisterSnapshotDataArgs,
+    ) -> impl 'agent + AsyncCall<Value = (ReadCanisterSnapshotDataResult,)> {
+        self.update(MgmtMethod::ReadCanisterSnapshotData.as_ref())
+            .with_arg(data_args)
+            .with_effective_canister_id(*canister_id)
+            .build()
+    }
+
+    /// Uploads the metadata of a canister snapshot by canister ID.
+    pub fn upload_canister_snapshot_metadata(
+        &self,
+        canister_id: &Principal,
+        metadata_args: &UploadCanisterSnapshotMetadataArgs,
+    ) -> impl 'agent + AsyncCall<Value = (UploadCanisterSnapshotMetadataResult,)> {
+        self.update(MgmtMethod::UploadCanisterSnapshotMetadata.as_ref())
+            .with_arg(metadata_args)
+            .with_effective_canister_id(*canister_id)
+            .build()
+    }
+
+    /// Uploads the data of a canister snapshot by canister ID and snapshot ID..
+    pub fn upload_canister_snapshot_data(
+        &self,
+        canister_id: &Principal,
+        data_args: &UploadCanisterSnapshotDataArgs,
+    ) -> impl 'agent + AsyncCall<Value = ()> {
+        self.update(MgmtMethod::UploadCanisterSnapshotData.as_ref())
+            .with_arg(data_args)
+            .with_effective_canister_id(*canister_id)
             .build()
     }
 }
