@@ -1,3 +1,4 @@
+#![cfg(unix)] // pocket-ic
 //! In this file, please mark all tests that require a running ic-ref as ignored.
 //!
 //! These tests are a Rust-like version using the Agent to cover the same tests
@@ -13,13 +14,13 @@
 //! use case being tested).
 use ref_tests::{universal_canister, with_agent};
 
-#[ignore]
-#[test]
-fn status_endpoint() {
-    with_agent(|agent| async move {
+#[tokio::test]
+async fn status_endpoint() {
+    with_agent(async move |_, agent| {
         agent.status().await?;
         Ok(())
     })
+    .await
 }
 
 mod management_canister {
@@ -34,9 +35,9 @@ mod management_canister {
         interfaces::{
             management_canister::{
                 builders::{
-                    CanisterSettings, CanisterUpgradeOptions, InstallMode, WasmMemoryPersistence,
+                    CanisterInstallMode, CanisterSettings, UpgradeFlags, WasmMemoryPersistence,
                 },
-                CanisterStatus, StatusCallResult,
+                CanisterStatusResult, CanisterStatusType,
             },
             wallet::CreateResult,
             ManagementCanister, WalletCanister,
@@ -65,27 +66,26 @@ mod management_canister {
         use ref_tests::get_effective_canister_id;
         use std::str::FromStr;
 
-        #[ignore]
-        #[test]
-        fn no_id_given() {
-            with_agent(|agent| async move {
+        #[tokio::test]
+        async fn no_id_given() {
+            with_agent(async move |pic, agent| {
                 let ic00 = ManagementCanister::create(&agent);
 
                 let _ = ic00
                     .create_canister()
                     .as_provisional_create_with_amount(None)
-                    .with_effective_canister_id(get_effective_canister_id().await)
+                    .with_effective_canister_id(get_effective_canister_id(pic).await)
                     .call_and_wait()
                     .await?;
 
                 Ok(())
             })
+            .await
         }
 
-        #[ignore]
-        #[test]
-        fn create_canister_necessary() {
-            with_agent(|agent| async move {
+        #[tokio::test]
+        async fn create_canister_necessary() {
+            with_agent(async move |_, agent| {
                 let ic00 = ManagementCanister::create(&agent);
                 let canister_wasm = b"\0asm\x01\0\0\0".to_vec();
 
@@ -106,39 +106,40 @@ mod management_canister {
                     error_code: Some(ref error_code)
                 }, .. }) if reject_message == "Canister 75hes-oqbaa-aaaaa-aaaaa-aaaaa-aaaaa-aaaaa-q not found" &&
                         error_code == "IC0301")
-                        || matches!(result, Err(HttpError(content)) if content.status == 400 && content.content == b"Canister 75hes-oqbaa-aaaaa-aaaaa-aaaaa-aaaaa-aaaaa-q does not belong to any subnet.")
+                        || matches!(result, Err(HttpError(content)) if content.status == 400
+                            && (content.content == b"Canister 75hes-oqbaa-aaaaa-aaaaa-aaaaa-aaaaa-aaaaa-q does not belong to any subnet."
+                                || content.content.starts_with(b"error: canister_not_found")))
                 );
 
                 Ok(())
-            })
+            }).await
         }
     }
 
-    #[ignore]
-    #[test]
-    fn management() {
+    #[tokio::test]
+    async fn management() {
         use ref_tests::get_effective_canister_id;
-        with_agent(|agent| async move {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
 
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .call_and_wait()
                 .await?;
             let canister_wasm = b"\0asm\x01\0\0\0".to_vec();
 
             // Install once.
             ic00.install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Install)
+                .with_mode(CanisterInstallMode::Install)
                 .call_and_wait()
                 .await?;
 
             // Re-install should fail.
             let result = ic00
                 .install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Install)
+                .with_mode(CanisterInstallMode::Install)
                 .call_and_wait()
                 .await;
 
@@ -146,28 +147,28 @@ mod management_canister {
 
             // Reinstall should succeed.
             ic00.install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Reinstall)
+                .with_mode(CanisterInstallMode::Reinstall)
                 .call_and_wait()
                 .await?;
 
             // Each agent has their own identity.
-            let other_agent_identity = create_basic_identity()?;
+            let other_agent_identity = create_basic_identity();
             let other_agent_principal = other_agent_identity.sender()?;
-            let other_agent = create_agent(other_agent_identity).await?;
+            let other_agent = create_agent(pic, other_agent_identity).await?;
             other_agent.fetch_root_key().await?;
             let other_ic00 = ManagementCanister::create(&other_agent);
 
             // Reinstall with another agent should fail.
             let result = other_ic00
                 .install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Reinstall)
+                .with_mode(CanisterInstallMode::Reinstall)
                 .call_and_wait()
                 .await;
             assert!(matches!(result, Err(AgentError::UncertifiedReject { .. })));
 
             // Upgrade should succeed.
             ic00.install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Upgrade(Some(CanisterUpgradeOptions {
+                .with_mode(CanisterInstallMode::Upgrade(Some(UpgradeFlags {
                     skip_pre_upgrade: Some(true),
                     wasm_memory_persistence: None,
                 })))
@@ -177,7 +178,7 @@ mod management_canister {
             // Upgrade with another agent should fail.
             let result = other_ic00
                 .install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Upgrade(Some(CanisterUpgradeOptions {
+                .with_mode(CanisterInstallMode::Upgrade(Some(UpgradeFlags {
                     skip_pre_upgrade: None,
                     wasm_memory_persistence: Some(WasmMemoryPersistence::Keep),
                 })))
@@ -202,14 +203,14 @@ mod management_canister {
                 reject_code: RejectCode::CanisterError,
                 reject_message,
                 error_code: Some(ref error_code),
-            }, .. }) if reject_message == format!("Only controllers of canister {} can call ic00 method update_settings", canister_id) &&
+            }, .. }) if reject_message == format!("Only controllers of canister {canister_id} can call ic00 method update_settings") &&
                     error_code == "IC0512")
             );
 
             // Reinstall as new controller
             other_ic00
                 .install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Reinstall)
+                .with_mode(CanisterInstallMode::Reinstall)
                 .call_and_wait()
                 .await?;
 
@@ -217,13 +218,13 @@ mod management_canister {
             let (canister_id_2,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .call_and_wait()
                 .await?;
 
             // Reinstall over empty canister
             ic00.install_code(&canister_id_2, &canister_wasm)
-                .with_mode(InstallMode::Reinstall)
+                .with_mode(CanisterInstallMode::Reinstall)
                 .call_and_wait()
                 .await?;
 
@@ -231,7 +232,7 @@ mod management_canister {
             let (canister_id_3,) = other_ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .call_and_wait()
                 .await?;
 
@@ -240,7 +241,7 @@ mod management_canister {
                 .canister_status(&canister_id_3)
                 .call_and_wait()
                 .await?;
-            assert_eq!(result.0.status, CanisterStatus::Running);
+            assert_eq!(result.0.status, CanisterStatusType::Running);
             assert_eq!(result.0.settings.controllers.len(), 1);
             assert_eq!(result.0.settings.controllers[0], other_agent_principal);
             assert_eq!(result.0.module_hash, None);
@@ -248,7 +249,7 @@ mod management_canister {
             // Install wasm.
             other_ic00
                 .install_code(&canister_id_3, &canister_wasm)
-                .with_mode(InstallMode::Install)
+                .with_mode(CanisterInstallMode::Install)
                 .call_and_wait()
                 .await?;
 
@@ -261,30 +262,29 @@ mod management_canister {
             assert_eq!(result.0.module_hash, Some(sha256_digest.to_vec()));
 
             Ok(())
-        })
+        }).await
     }
 
-    #[ignore]
-    #[test]
-    fn multiple_canisters_aaaaa_aa_but_really_provisional() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn multiple_canisters_aaaaa_aa_but_really_provisional() {
+        with_agent(async move |pic, agent| {
             let agent_principal = agent.get_principal()?;
             // Each agent has their own identity.
-            let other_agent_identity = create_basic_identity()?;
+            let other_agent_identity = create_basic_identity();
             let other_agent_principal = other_agent_identity.sender()?;
-            let other_agent = create_agent(other_agent_identity).await?;
+            let other_agent = create_agent(pic, other_agent_identity).await?;
             other_agent.fetch_root_key().await?;
             let other_ic00 = ManagementCanister::create(&other_agent);
 
             let secp256k1_identity = create_secp256k1_identity()?;
             let secp256k1_principal = secp256k1_identity.sender()?;
-            let secp256k1_agent = create_agent(secp256k1_identity).await?;
+            let secp256k1_agent = create_agent(pic, secp256k1_identity).await?;
             secp256k1_agent.fetch_root_key().await?;
             let secp256k1_ic00 = ManagementCanister::create(&secp256k1_agent);
 
             let prime256v1_identity = create_prime256v1_identity()?;
             let prime256v1_principal = prime256v1_identity.sender()?;
-            let prime256v1_agent = create_agent(prime256v1_identity).await?;
+            let prime256v1_agent = create_agent(pic, prime256v1_identity).await?;
             prime256v1_agent.fetch_root_key().await?;
             let prime256v1_ic00 = ManagementCanister::create(&prime256v1_agent);
 
@@ -293,7 +293,7 @@ mod management_canister {
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None) // ok
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 //.with_canister_id("aaaaa-aa")
                 .with_controller(agent_principal)
                 .with_controller(other_agent_principal)
@@ -404,6 +404,7 @@ mod management_canister {
 
             Ok(())
         })
+        .await
     }
 
     fn assert_err_or_reject<S>(
@@ -423,40 +424,38 @@ mod management_canister {
 
         assert!(
             matches!(result, Err(AgentError::HttpError(_))),
-            "expect an HttpError, or a CertifiedReject with reject_code in {:?}",
-            allowed_reject_codes
+            "expect an HttpError, or a CertifiedReject with reject_code in {allowed_reject_codes:?}"
         );
     }
 
-    #[ignore]
-    #[test]
-    fn canister_lifecycle_and_delete() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn canister_lifecycle_and_delete() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .call_and_wait()
                 .await?;
             let canister_wasm = b"\0asm\x01\0\0\0".to_vec();
 
             // Install once.
             ic00.install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Install)
+                .with_mode(CanisterInstallMode::Install)
                 .call_and_wait()
                 .await?;
 
             // A newly installed canister should be running
             let result = ic00.canister_status(&canister_id).call_and_wait().await;
-            assert_eq!(result?.0.status, CanisterStatus::Running);
+            assert_eq!(result?.0.status, CanisterStatusType::Running);
 
             // Stop should succeed.
             ic00.stop_canister(&canister_id).call_and_wait().await?;
 
             // Canister should be stopped
             let result = ic00.canister_status(&canister_id).call_and_wait().await;
-            assert_eq!(result?.0.status, CanisterStatus::Stopped);
+            assert_eq!(result?.0.status, CanisterStatusType::Stopped);
 
             // Another stop is a noop
             ic00.stop_canister(&canister_id).call_and_wait().await?;
@@ -485,7 +484,7 @@ mod management_canister {
                         reject_code: RejectCode::CanisterError,
                         reject_message,
                         error_code: Some(error_code),
-                    }, .. }) if *reject_message == format!("IC0508: Canister {canister_id} is stopped and therefore does not have a CallContextManager")
+                    }, .. }) if *reject_message == format!("Canister {canister_id} is stopped and therefore does not have a CallContextManager")
                         && error_code == "IC0508"
                 ),
                 "wrong error: {result:?}"
@@ -493,7 +492,7 @@ mod management_canister {
 
             // Upgrade should succeed
             ic00.install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Upgrade(Some(CanisterUpgradeOptions {
+                .with_mode(CanisterInstallMode::Upgrade(Some(UpgradeFlags {
                     skip_pre_upgrade: None,
                     wasm_memory_persistence: Some(WasmMemoryPersistence::Replace),
                 })))
@@ -505,7 +504,7 @@ mod management_canister {
 
             // Canister should be running
             let result = ic00.canister_status(&canister_id).call_and_wait().await;
-            assert_eq!(result?.0.status, CanisterStatus::Running);
+            assert_eq!(result?.0.status, CanisterStatusType::Running);
 
             // Can call update
             let result = agent.update(&canister_id, "update").call_and_wait().await;
@@ -531,7 +530,7 @@ mod management_canister {
                         reject_code: RejectCode::CanisterError,
                         reject_message,
                         error_code: Some(error_code),
-                    }, .. }) if reject_message.contains(&format!("Canister {}: Canister has no query method 'query'", canister_id))
+                    }, .. }) if reject_message.contains(&format!("Canister {canister_id}: Canister has no query method 'query'"))
                         && error_code == "IC0536",
                 ),
                 "wrong error: {result:?}"
@@ -555,7 +554,7 @@ mod management_canister {
                         reject_code: RejectCode::DestinationInvalid,
                         reject_message,
                         error_code: Some(error_code),
-                    }, .. }) if *reject_message == format!("Canister {} not found", canister_id)
+                    }, .. }) if *reject_message == format!("Canister {canister_id} not found")
                         && error_code == "IC0301"
                 ),
                 "wrong error: {result:?}"
@@ -570,7 +569,7 @@ mod management_canister {
                         reject_code: RejectCode::DestinationInvalid,
                         reject_message,
                         error_code: Some(error_code),
-                    }, .. }) if *reject_message == format!("IC0301: Canister {} not found", canister_id)
+                    }, .. }) if *reject_message == format!("Canister {canister_id} not found")
                         && error_code == "IC0301"
                 ),
                 "wrong error: {result:?}"
@@ -588,7 +587,7 @@ mod management_canister {
                                 error_code: Some(error_code),
                             },
                         ..
-                    }) if *reject_message == format!("Canister {} not found", canister_id)
+                    }) if *reject_message == format!("Canister {canister_id} not found")
                         && error_code == "IC0301" =>
                     {
                         true
@@ -608,37 +607,36 @@ mod management_canister {
                         reject_code: RejectCode::DestinationInvalid,
                         reject_message,
                         error_code: Some(error_code),
-                    }, .. }) if *reject_message == format!("Canister {} not found", canister_id)
+                    }, .. }) if *reject_message == format!("Canister {canister_id} not found")
                         && error_code == "IC0301"
                 ),
                 "wrong error: {result:?}"
             );
             Ok(())
-        })
+        }).await
     }
 
-    #[ignore]
-    #[test]
-    fn canister_lifecycle_as_wrong_controller() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn canister_lifecycle_as_wrong_controller() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .call_and_wait()
                 .await?;
             let canister_wasm = b"\0asm\x01\0\0\0".to_vec();
 
             // Install once.
             ic00.install_code(&canister_id, &canister_wasm)
-                .with_mode(InstallMode::Install)
+                .with_mode(CanisterInstallMode::Install)
                 .call_and_wait()
                 .await?;
 
             // Create another agent with different identity.
-            let other_agent_identity = create_basic_identity()?;
-            let other_agent = create_agent(other_agent_identity).await?;
+            let other_agent_identity = create_basic_identity();
+            let other_agent = create_agent(pic, other_agent_identity).await?;
             other_agent.fetch_root_key().await?;
             let other_ic00 = ManagementCanister::create(&other_agent);
 
@@ -654,7 +652,7 @@ mod management_canister {
                         reject_code: RejectCode::CanisterError,
                         reject_message,
                         error_code: Some(error_code),
-                    }, .. }) if *reject_message == format!("Only controllers of canister {} can call ic00 method start_canister", canister_id)
+                    }, .. }) if *reject_message == format!("Only controllers of canister {canister_id} can call ic00 method start_canister")
                         && error_code == "IC0512"
                 ),
                 "wrong error: {result:?}"
@@ -669,7 +667,7 @@ mod management_canister {
                         reject_code: RejectCode::CanisterError,
                         reject_message,
                         error_code: Some(error_code),
-                    }, ..}) if *reject_message == format!("Only controllers of canister {} can call ic00 method stop_canister", canister_id)
+                    }, ..}) if *reject_message == format!("Only controllers of canister {canister_id} can call ic00 method stop_canister")
                         && error_code == "IC0512"
                 ),
                 "wrong error: {result:?}"
@@ -712,13 +710,12 @@ mod management_canister {
             );
 
             Ok(())
-        })
+        }).await
     }
 
-    #[ignore]
-    #[test]
-    fn provisional_create_canister_with_cycles() {
-        with_wallet_canister(None, |agent, wallet_id| async move {
+    #[tokio::test]
+    async fn provisional_create_canister_with_cycles() {
+        with_wallet_canister(None, async move |pic, agent, wallet_id| {
             let default_canister_balance: u128 = 100_000_000_000_000;
 
             // empty cycle balance on create
@@ -740,6 +737,7 @@ mod management_canister {
                     wasm_memory_limit: None,
                     wasm_memory_threshold: None,
                     log_visibility: None,
+                    environment_variables: None,
                 },
             };
 
@@ -765,7 +763,7 @@ mod management_canister {
             let status_args = In { canister_id };
             let args = Argument::from_candid((status_args,));
 
-            let (result,): (StatusCallResult,) = wallet
+            let (result,): (CanisterStatusResult,) = wallet
                 .call(Principal::management_canister(), "canister_status", args, 0)
                 .call_and_wait()
                 .await?;
@@ -782,7 +780,7 @@ mod management_canister {
             let (canister_id_1,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .call_and_wait()
                 .await?;
             let result = ic00.canister_status(&canister_id_1).call_and_wait().await?;
@@ -797,7 +795,7 @@ mod management_canister {
             let (canister_id_2,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(Some(amount))
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .call_and_wait()
                 .await?;
             let result = ic00.canister_status(&canister_id_2).call_and_wait().await?;
@@ -810,12 +808,12 @@ mod management_canister {
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn randomness() {
-        with_wallet_canister(None, |agent, wallet_id| async move {
+    #[tokio::test]
+    async fn randomness() {
+        with_wallet_canister(None, async move |_, agent, wallet_id| {
             let wallet = WalletCanister::create(&agent, wallet_id).await?;
             let (rand_1,): (Vec<u8>,) = wallet
                 .call(
@@ -855,27 +853,40 @@ mod management_canister {
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn chunked_wasm() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn chunked_wasm() {
+        use ic_management_canister_types::UploadChunkArgs;
+        with_agent(async move |pic, agent| {
             let wasm = b"\0asm\x01\0\0\0";
             let wasm_hash = Sha256::digest(wasm).to_vec();
             let mgmt = ManagementCanister::create(&agent);
             let (canister_id,) = mgmt
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .call_and_wait()
                 .await?;
             let (pt1,) = mgmt
-                .upload_chunk(&canister_id, &wasm[0..4])
+                .upload_chunk(
+                    &canister_id,
+                    &UploadChunkArgs {
+                        canister_id,
+                        chunk: wasm[0..4].to_vec(),
+                    },
+                )
                 .call_and_wait()
                 .await?;
             let (pt2,) = mgmt
-                .upload_chunk(&canister_id, &wasm[4..8])
+                .upload_chunk(
+                    &canister_id,
+                    &UploadChunkArgs {
+                        canister_id,
+                        chunk: wasm[4..8].to_vec(),
+                    },
+                )
                 .call_and_wait()
                 .await?;
             mgmt.install_chunked_code(&canister_id, &wasm_hash)
@@ -884,24 +895,24 @@ mod management_canister {
                 .await?;
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
+    #[tokio::test]
     // makes sure that calling fetch_root_key twice by accident does not break
-    fn multi_fetch_root_key() {
-        with_agent(|agent| async move {
+    async fn multi_fetch_root_key() {
+        with_agent(async move |_, agent| {
             agent.fetch_root_key().await?;
             agent.fetch_root_key().await?;
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn subnet_metrics() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn subnet_metrics() {
+        with_agent(async move |_, agent| {
             let ic00 = ManagementCanister::create(&agent);
 
             // create a canister on the root subnet
@@ -923,6 +934,25 @@ mod management_canister {
             );
             Ok(())
         })
+        .await
+    }
+
+    #[tokio::test]
+    async fn subnet_canister_ranges() {
+        with_agent(async move |_, agent| {
+            // fetch root subnet canister ranges
+            let ranges = agent
+                .read_state_subnet_canister_ranges(Principal::self_authenticating(
+                    agent.read_root_key(),
+                ))
+                .await?;
+            assert!(
+                !ranges.is_empty(),
+                "expected at least one canister range on the root subnet"
+            );
+            Ok(())
+        })
+        .await
     }
 }
 
@@ -934,10 +964,9 @@ mod simple_calls {
     };
     use ref_tests::with_universal_canister;
 
-    #[ignore]
-    #[test]
-    fn call() {
-        with_universal_canister(|agent, canister_id| async move {
+    #[tokio::test]
+    async fn call() {
+        with_universal_canister(async move |_, agent, canister_id| {
             let arg = payload().reply_data(b"hello").build();
             let result = agent
                 .update(&canister_id, "update")
@@ -948,12 +977,12 @@ mod simple_calls {
             assert_eq!(result.as_slice(), b"hello");
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn query() {
-        with_universal_canister(|agent, canister_id| async move {
+    #[tokio::test]
+    async fn query() {
+        with_universal_canister(async move |_, agent, canister_id| {
             let arg = payload().reply_data(b"hello").build();
             let result = agent
                 .query(&canister_id, "query")
@@ -973,12 +1002,12 @@ mod simple_calls {
             assert_eq!(result, b"hello");
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn non_existant_call() {
-        with_universal_canister(|agent, canister_id| async move {
+    #[tokio::test]
+    async fn non_existant_call() {
+        with_universal_canister(async move |_, agent, canister_id| {
             let arg = payload().reply_data(b"hello").build();
             let result = agent
                 .update(&canister_id, "non_existent_method")
@@ -1001,12 +1030,12 @@ mod simple_calls {
             );
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn non_existant_query() {
-        with_universal_canister(|agent, canister_id| async move {
+    #[tokio::test]
+    async fn non_existant_query() {
+        with_universal_canister(async move |_, agent, canister_id| {
             let arg = payload().reply_data(b"hello").build();
             let result = agent
                 .query(&canister_id, "non_existent_method")
@@ -1029,6 +1058,7 @@ mod simple_calls {
             );
             Ok(())
         })
+        .await
     }
 }
 
@@ -1048,16 +1078,15 @@ mod extras {
     };
     use ref_tests::{get_effective_canister_id, with_agent};
 
-    #[ignore]
-    #[test]
-    fn valid_allocations() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn valid_allocations() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
 
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(Some(20_000_000_000_000_u128))
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_compute_allocation(1_u64)
                 .with_memory_allocation(1024 * 1024_u64)
                 .with_freezing_threshold(1_000_000_u64)
@@ -1077,23 +1106,23 @@ mod extras {
             );
             assert_eq!(
                 result.0.settings.reserved_cycles_limit,
-                Some(Nat::from(2_500_800_000_000u128))
+                Nat::from(2_500_800_000_000u128)
             );
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn memory_allocation() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn memory_allocation() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
             // Prevent creating with over 1 << 48. This does not contact the server.
             let result = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_memory_allocation(1u64 << 50)
                 .call_and_wait()
                 .await;
@@ -1105,45 +1134,45 @@ mod extras {
             let (_,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_memory_allocation(10 * 1024 * 1024u64)
                 .call_and_wait()
                 .await?;
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn compute_allocation() {
+    #[tokio::test]
+    async fn compute_allocation() {
         use std::convert::TryFrom;
 
-        with_agent(|agent| async move {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
             let ca = ComputeAllocation::try_from(10).unwrap();
 
             let (_,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(Some(1_000_000_000_000_000))
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_compute_allocation(ca)
                 .call_and_wait()
                 .await?;
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn freezing_threshold() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn freezing_threshold() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
             let result = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_freezing_threshold(2u128.pow(70))
                 .call_and_wait()
                 .await;
@@ -1154,18 +1183,18 @@ mod extras {
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn create_with_reserved_cycles_limit() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn create_with_reserved_cycles_limit() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
 
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_reserved_cycles_limit(2u128.pow(70))
                 .call_and_wait()
                 .await
@@ -1174,23 +1203,23 @@ mod extras {
             let result = ic00.canister_status(&canister_id).call_and_wait().await?;
             assert_eq!(
                 result.0.settings.reserved_cycles_limit,
-                Some(Nat::from(2u128.pow(70)))
+                Nat::from(2u128.pow(70))
             );
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn update_reserved_cycles_limit() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn update_reserved_cycles_limit() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
 
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(Some(20_000_000_000_000_u128))
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_reserved_cycles_limit(2_500_800_000_000u128)
                 .call_and_wait()
                 .await?;
@@ -1198,7 +1227,7 @@ mod extras {
             let result = ic00.canister_status(&canister_id).call_and_wait().await?;
             assert_eq!(
                 result.0.settings.reserved_cycles_limit,
-                Some(Nat::from(2_500_800_000_000u128))
+                Nat::from(2_500_800_000_000u128)
             );
 
             ic00.update_settings(&canister_id)
@@ -1209,7 +1238,7 @@ mod extras {
             let result = ic00.canister_status(&canister_id).call_and_wait().await?;
             assert_eq!(
                 result.0.settings.reserved_cycles_limit,
-                Some(Nat::from(3_400_200_000_000u128))
+                Nat::from(3_400_200_000_000u128)
             );
 
             let no_change: Option<u128> = None;
@@ -1221,17 +1250,17 @@ mod extras {
             let result = ic00.canister_status(&canister_id).call_and_wait().await?;
             assert_eq!(
                 result.0.settings.reserved_cycles_limit,
-                Some(Nat::from(3_400_200_000_000u128))
+                Nat::from(3_400_200_000_000u128)
             );
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn specified_id() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn specified_id() {
+        with_agent(async move |_, agent| {
             let ic00 = ManagementCanister::create(&agent);
             let specified_id = Principal::from_text("iimsn-6yaaa-aaaaa-afiaa-cai").unwrap(); // [42, 0] should be large enough
             assert_eq!(
@@ -1276,18 +1305,18 @@ mod extras {
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn create_with_wasm_memory_limit() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn create_with_wasm_memory_limit() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
 
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_wasm_memory_limit(1_000_000_000)
                 .call_and_wait()
                 .await
@@ -1296,23 +1325,23 @@ mod extras {
             let result = ic00.canister_status(&canister_id).call_and_wait().await?;
             assert_eq!(
                 result.0.settings.wasm_memory_limit,
-                Some(Nat::from(1_000_000_000_u64))
+                Nat::from(1_000_000_000_u64)
             );
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn update_wasm_memory_limit() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn update_wasm_memory_limit() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
 
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(Some(20_000_000_000_000_u128))
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_wasm_memory_limit(1_000_000_000)
                 .call_and_wait()
                 .await?;
@@ -1320,7 +1349,7 @@ mod extras {
             let result = ic00.canister_status(&canister_id).call_and_wait().await?;
             assert_eq!(
                 result.0.settings.wasm_memory_limit,
-                Some(Nat::from(1_000_000_000_u64))
+                Nat::from(1_000_000_000_u64)
             );
 
             ic00.update_settings(&canister_id)
@@ -1331,7 +1360,7 @@ mod extras {
             let result = ic00.canister_status(&canister_id).call_and_wait().await?;
             assert_eq!(
                 result.0.settings.wasm_memory_limit,
-                Some(Nat::from(3_000_000_000_u64))
+                Nat::from(3_000_000_000_u64)
             );
 
             let no_change: Option<u64> = None;
@@ -1343,23 +1372,23 @@ mod extras {
             let result = ic00.canister_status(&canister_id).call_and_wait().await?;
             assert_eq!(
                 result.0.settings.wasm_memory_limit,
-                Some(Nat::from(3_000_000_000_u64))
+                Nat::from(3_000_000_000_u64)
             );
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn create_with_log_visibility() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn create_with_log_visibility() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
 
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(None)
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_log_visibility(LogVisibility::Public)
                 .call_and_wait()
                 .await
@@ -1370,19 +1399,19 @@ mod extras {
 
             Ok(())
         })
+        .await
     }
 
-    #[ignore]
-    #[test]
-    fn update_log_visibility() {
-        with_agent(|agent| async move {
+    #[tokio::test]
+    async fn update_log_visibility() {
+        with_agent(async move |pic, agent| {
             let ic00 = ManagementCanister::create(&agent);
 
             // Create with Controllers.
             let (canister_id,) = ic00
                 .create_canister()
                 .as_provisional_create_with_amount(Some(20_000_000_000_000_u128))
-                .with_effective_canister_id(get_effective_canister_id().await)
+                .with_effective_canister_id(get_effective_canister_id(pic).await)
                 .with_log_visibility(LogVisibility::Controllers)
                 .call_and_wait()
                 .await?;
@@ -1411,5 +1440,6 @@ mod extras {
 
             Ok(())
         })
+        .await
     }
 }
